+++ webservice/src/api/administration.rs
@@ -3,10 +3,7 @@ use crate::{
     actions, errors::ServiceError, extractors::ClaimsFromAuth, settings::Settings, DbPool,
 };
 use actix_files::NamedFile;
-use actix_web::{
-    web::{self, HttpResponse},
-    HttpRequest,
-};
+use actix_web::{web, HttpRequest, HttpResponse};
 use futures::prelude::*;
 use graphannis::CorpusStorage;
 use std::io::Seek;
@@ -52,11 +49,11 @@ pub async fn list_groups(
     check_is_admin(&claims.0)?;
 
     let conn = db_pool.get()?;
-    let corpus_groups = web::block::<_, _, ServiceError>(move || {
+    let corpus_groups = web::block::<_, Result<_, ServiceError>>(move || {
         let result = actions::list_groups(&conn)?;
         Ok(result)
     })
-    .await?;
+    .await??;
 
     Ok(HttpResponse::Ok().json(corpus_groups))
 }
@@ -69,7 +66,8 @@ pub async fn delete_group(
     check_is_admin(&claims.0)?;
 
     let conn = db_pool.get()?;
-    web::block::<_, _, ServiceError>(move || actions::delete_group(&group_name, &conn)).await?;
+    web::block::<_, Result<_, ServiceError>>(move || actions::delete_group(&group_name, &conn))
+        .await??;
 
     Ok(HttpResponse::Ok().json("Group deleted"))
 }
@@ -87,8 +85,10 @@ pub async fn put_group(
     }
 
     let conn = db_pool.get()?;
-    web::block::<_, _, ServiceError>(move || actions::add_or_replace_group(group.clone(), &conn))
-        .await?;
+    web::block::<_, Result<_, ServiceError>>(move || {
+        actions::add_or_replace_group(group.clone(), &conn)
+    })
+    .await??;
 
     Ok(HttpResponse::Ok().json("Group added/replaced"))
 }
@@ -118,7 +118,7 @@ pub async fn import_corpus(
     let mut tmp = tempfile::tempfile()?;
     while let Some(chunk) = body.next().await {
         let data = chunk?;
-        tmp = web::block(move || tmp.write_all(&data).map(|_| tmp)).await?;
+        tmp = web::block(move || tmp.write_all(&data).map(|_| tmp)).await??;
     }
 
     // Create a UUID which is used for the background job
@@ -289,7 +289,7 @@ pub async fn jobs(
             JobStatus::Finished(result) => {
                 if let Some((tmp_file, file_name)) = result {
                     let named_file = NamedFile::from_file(tmp_file, file_name)?;
-                    let response = named_file.into_response(&req)?;
+                    let response = named_file.into_response(&req);
                     return Ok(response);
                 } else {
                     return Ok(HttpResponse::Ok().json(j.messages));